fix: 字段类型

+++ backend/init.sql
@@ -52,7 +52,7 @@ CREATE TABLE video_tanmu
     from_name        VARCHAR(255) DEFAULT '播放地址1' COMMENT '播放来源',
     play_name        VARCHAR(255) DEFAULT '第一集' COMMENT '选集',
     play_url         VARCHAR(255) DEFAULT '' COMMENT '播放url',
-    `current_time`   FLOAT(5, 2)  DEFAULT 0 COMMENT '弹幕时间',
+    `current_time`   FLOAT(28, 2)  DEFAULT 0 COMMENT '弹幕时间',
     current_time_int INT(11)      DEFAULT 0 COMMENT '时间取整数',
     content          VARCHAR(255) DEFAULT '' COMMENT '弹幕内容',
     style            VARCHAR(255) DEFAULT '' COMMENT '样式'
